# contracts/web-console.openapi.json
{
  "components": {
    "schemas": {
...
      "AnomalyResponse": {
        "additionalProperties": true,
        "properties": {
          "anomaly_type": {
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "event_time": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "observed_value": {
            "nullable": true,
            "type": "number"
          },
          "severity": {
            "enum": [
              "info",
              "warning",
              "critical"
            ],
            "type": "string"
          },
          "symbol": {
            "type": "string"
          },
          "threshold_value": {
            "nullable": true,
            "type": "number"
          }
        },
        "required": [
          "id",
          "symbol",
          "anomaly_type",
          "severity",
          "message",
          "observed_value",
          "threshold_value",
          "event_time",
          "created_at"
        ],
        "type": "object"
      },
      "DashboardHealthSummary": {
        "additionalProperties": true,
        "properties": {
          "evaluated_at": {
            "format": "date-time",
            "type": "string"
          },
          "recent_anomaly_count": {
            "type": "integer"
          },
          "score": {
            "maximum": 100,
            "minimum": 0,
            "type": "integer"
          },
          "status": {
            "enum": [
              "healthy",
              "degraded",
              "unhealthy"
            ],
            "type": "string"
          }
        },
        "required": [
          "score",
          "status",
          "recent_anomaly_count",
          "evaluated_at"
        ],
        "type": "object"
      },
      "DashboardStateSummary": {
        "additionalProperties": true,
        "properties": {
          "best_ask_price": {
            "description": "rust_decimal JSON serialization",
            "nullable": true,
            "type": "string"
          },
          "best_bid_price": {
            "description": "rust_decimal JSON serialization",
            "nullable": true,
            "type": "string"
          },
          "depth_sequence_gap_count": {
            "type": "integer"
          },
          "last_event_age_ms": {
            "nullable": true,
            "type": "integer"
          },
          "last_event_time": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "last_trade_price": {
            "description": "rust_decimal JSON serialization",
            "nullable": true,
            "type": "string"
          },
          "price_change_1m_pct": {
            "nullable": true,
            "type": "number"
          },
          "spread_pct": {
            "nullable": true,
            "type": "number"
          },
          "trades_per_minute": {
            "nullable": true,
            "type": "number"
          }
        },
        "required": [
          "last_trade_price",
          "best_bid_price",
          "best_ask_price",
          "spread_pct",
          "price_change_1m_pct",
          "trades_per_minute",
          "last_event_time",
          "last_event_age_ms",
          "depth_sequence_gap_count"
        ],
        "type": "object"
      },
      "DashboardSummaryResponse": {
        "additionalProperties": true,
        "properties": {
          "pipeline": {
            "$ref": "#/components/schemas/PipelineHealthResponse"
          },
          "recent_anomalies": {
            "items": {
              "$ref": "#/components/schemas/AnomalyResponse"
            },
            "type": "array"
          },
          "service": {
            "$ref": "#/components/schemas/HealthResponse"
          },
          "symbols": {
            "items": {
              "$ref": "#/components/schemas/DashboardSymbolSummary"
            },
            "type": "array"
          }
        },
        "required": [
          "service",
          "pipeline",
          "symbols",
          "recent_anomalies"
        ],
        "type": "object"
      },
      "DashboardSymbolSummary": {
        "additionalProperties": true,
        "properties": {
          "health": {
            "$ref": "#/components/schemas/DashboardHealthSummary",
            "nullable": true
          },
          "state": {
            "$ref": "#/components/schemas/DashboardStateSummary",
            "nullable": true
          },
          "symbol": {
            "type": "string"
          }
        },
        "required": [
          "symbol",
          "state",
          "health"
        ],
        "type": "object"
      },
...
      "HealthResponse": {
        "additionalProperties": true,
        "properties": {
          "service": {
            "enum": [
              "signalguard-rs"
            ],
            "type": "string"
          },
          "status": {
            "enum": [
              "ok"
            ],
            "type": "string"
          }
        },
        "required": [
          "status",
          "service"
        ],
        "type": "object"
      },
...
      "PipelineHealthResponse": {
        "additionalProperties": true,
        "properties": {
          "cache_errors": {
            "type": "integer"
          },
          "last_message_age_ms": {
            "nullable": true,
            "type": "integer"
          },
          "parse_errors": {
            "type": "integer"
          },
          "reconnect_attempts": {
            "type": "integer"
          },
          "status": {
            "enum": [
              "healthy",
              "degraded",
              "unhealthy"
            ],
            "type": "string"
          },
          "storage_errors": {
            "type": "integer"
          }
        },
        "required": [
          "status",
          "last_message_age_ms",
          "parse_errors",
          "reconnect_attempts",
          "storage_errors",
          "cache_errors"
        ],
        "type": "object"
      },
      "RuntimeModeResponse": {
        "additionalProperties": true,
        "properties": {
          "last_error": {
            "nullable": true,
            "type": "string"
          },
          "last_started_at": {
            "format": "date-time",
            "type": "string"
          },
          "last_switched_at": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "mode": {
            "enum": [
              "replay",
              "live"
            ],
            "type": "string"
          },
          "mode_label": {
            "type": "string"
          },
          "source": {
            "enum": [
              "config",
              "runtime"
            ],
            "type": "string"
          },
          "status": {
            "enum": [
              "starting",
              "running",
              "switching",
              "failed",
              "stopped",
              "completed"
            ],
            "type": "string"
          },
          "switching_supported": {
            "type": "boolean"
          },
          "symbols": {
            "items": {
              "type": "string"
            },
            "type": "array"
          }
        },
        "required": [
          "mode",
          "mode_label",
          "status",
          "symbols",
          "switching_supported",
          "source",
          "last_started_at",
          "last_switched_at",
          "last_error"
        ],
        "type": "object"
      },
      "RuntimeModeSwitchRequest": {
        "additionalProperties": true,
        "properties": {
          "mode": {
            "enum": [
              "replay",
              "live"
            ],
            "type": "string"
          },
          "reset_state": {
            "type": "boolean"
          },
          "reset_storage": {
            "type": "boolean"
          },
          "symbols": {
            "items": {
              "type": "string"
            },
            "type": "array"
          }
        },
        "required": [
          "mode"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "SignalGuard web console API",
    "version": "0.4.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/dashboard/summary": {
      "get": {
        "operationId": "getDashboardSummary",
        "parameters": [
          {
            "in": "query",
            "name": "mode",
            "required": false,
            "schema": {
              "enum": [
                "demo",
                "live"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardSummaryResponse"
                }
              }
            },
            "description": "OK"
          }
        }
      }
    },
...
    "/runtime/mode": {
      "get": {
        "operationId": "getRuntimeMode",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeModeResponse"
                }
              }
            },
            "description": "OK"
          }
        }
      },
      "post": {
        "operationId": "postRuntimeMode",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuntimeModeSwitchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeModeResponse"
                }
              }
            },
            "description": "OK"
          }
        }
      }
    },
...
  },
...
}